inserePedido funcional e removePedido 'funcional'

+++ src/main/java/dcc025/ufjf/trabalho/Estoque.java
@@ -75,7 +75,7 @@ public class  Estoque
         }
     }
     
-    public void remEstoque(String ingredienteNome, int quantidade){
+    public void remEstoque(String ingredienteNome, float quantidade){
         
         ItemEstoque ingrediente = new ItemEstoque(ingredienteNome, quantidade, " ");
